simple start (adds WorldModule)
