+++ src/main/java/org/skriptlang/skript/bukkit/BukkitModule.java
@@ -24,7 +24,7 @@ import org.skriptlang.skript.bukkit.potion.PotionModule;
 import org.skriptlang.skript.bukkit.tags.TagModule;
 import org.skriptlang.skript.bukkit.text.TextModule;
 import org.skriptlang.skript.bukkit.types.*;
-import org.skriptlang.skript.bukkit.worldborder.elements.WorldBorderModule;
+import org.skriptlang.skript.bukkit.world.WorldModule;
 
 import java.util.List;
 
@@ -56,7 +56,7 @@ public class BukkitModule extends HierarchicalAddonModule {
 			new PotionModule(this),
 			new TagModule(this),
 			new TextModule(this),
-			new WorldBorderModule(this)
+			new WorldModule(this)
 		);
 	}
