+++ src/main/java/org/skriptlang/skript/bukkit/world/worldborder/elements/expressions/ExprWorldBorder.java
@@ -1,4 +1,4 @@
-package org.skriptlang.skript.bukkit.worldborder.elements.expressions;
+package org.skriptlang.skript.bukkit.world.worldborder.elements.expressions;
 
 import ch.njol.skript.Skript;
 import ch.njol.skript.classes.Changer.ChangeMode;
